feat: 회원가입, 로그인 부분 추가, security 적용, 프로젝트 구조에 맞게 클래스 위치 변경, service, r…

+++ src/main/java/com/project/eatTogether/application/dto/FilterDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class FilterDTO {
     private Long rsId;            // 식당 ID
     private String rsName;        // 식당 이름
-    private int rsAvgRate;        // 평균 평점
+    private Double rsAvgRate;        // 평균 평점
     private int rsReviewCount;    // 리뷰 수
     private int rsBookmarkCount;  // 북마크 수
 }
